separate calling command to hla and virus
