+++ src/calling/haplotypes/mod.rs
@@ -1,0 +1,3 @@
+pub mod haplotypes;
+pub mod hla;
+pub mod virus;
